Extended MessageRequest paramters and poll handling
Added MessageRequest parameters:
* last_delivered_at
* acknowledged_by_device
* acknowledged_by

Added convenience functions for poll handling
* has_expired()
* was_called_back()
* was_acknowledged()
* has_ended()

Before this change polling an acknowledged, expired or called_back
priority-2-notofication updated the MessageRequest parameters
only on the first call and returned None without further contacting the
pushover servers on the following calls to poll(). So not only did
it take at least two iterations in the "while request.poll():" loop, but
also missed the fact that in the pushover api a priority-2-message
can first expire and then still get acknowledged afterwards.
This was fixed as well.